Skyline/Crawdad: add a cast on return value for quiet compile.  No functional change.
Former-commit-id: e6a36dae549da838164017cbaff57558bde696ab

+++ pwiz_tools/Shared/Crawdad/ManagedCrawdad.cpp
@@ -48,7 +48,7 @@ namespace Crawdad {
                 _CrtSetReportMode(_CRT_WARN, _CRTDBG_MODE_DEBUG);
                 _CrtMemDumpAllObjectsSince(&_checkPointState);
             }
-            return diffState.lSizes[_NORMAL_BLOCK];
+            return (long)diffState.lSizes[_NORMAL_BLOCK];
         }
 #endif
         return 0;
